requirements.txt
tensorflow==2.15.0
keras==2.15.0
pillow>=9.0.0
numpy>=1.21